+++ src/main/java/net/glowstone/net/codec/play/entity/SpawnXpOrbCodec.java
@@ -11,9 +11,9 @@ public final class SpawnXpOrbCodec implements Codec<SpawnXpOrbMessage> {
     @Override
     public SpawnXpOrbMessage decode(ByteBuf buf) throws IOException {
         int id = ByteBufUtils.readVarInt(buf);
-        int x = buf.readInt();
-        int y = buf.readInt();
-        int z = buf.readInt();
+        double x = buf.readDouble();
+        double y = buf.readDouble();
+        double z = buf.readDouble();
         short count = buf.readShort();
         return new SpawnXpOrbMessage(id, x, y, z, count);
     }
@@ -21,9 +21,9 @@ public final class SpawnXpOrbCodec implements Codec<SpawnXpOrbMessage> {
     @Override
     public ByteBuf encode(ByteBuf buf, SpawnXpOrbMessage message) throws IOException {
         ByteBufUtils.writeVarInt(buf, message.getId());
-        buf.writeInt(message.getX());
-        buf.writeInt(message.getY());
-        buf.writeInt(message.getZ());
+        buf.writeDouble(message.getX());
+        buf.writeDouble(message.getY());
+        buf.writeDouble(message.getZ());
         buf.writeShort(message.getCount());
         return buf;
     }
